Split video file repository
